Change LINKFLAGS in configure checks for boost_python.
When boost_python static library is not available and there is only
shared version, link with the Python shared library so that scons
conf.CheckLib does not fail due to unresolved Python symbols.

A proper way might be to configure boost library so that the static
library gets built anyway.